♻️ refactor: refactor mobile layout
