+++ package.json
@@ -80,7 +80,7 @@
   "dependencies": {
     "@ant-design/icons": "^5",
     "@anthropic-ai/sdk": "^0.18.0",
-    "@auth/core": "latest",
+    "@auth/core": "0.28.0",
     "@aws-sdk/client-bedrock-runtime": "^3.525.0",
     "@azure/openai": "^1.0.0-beta.11",
     "@cfworker/json-schema": "^1",
