+++ src/app/chat/_layout/Desktop/SessionHeader.tsx
